refactor(): revert time picker column change
